remove leaderbord component, move content to files in app
